The problem fixed here is a result of allowing numbers to be used as if they were names - an amibuguity results for the expression "(a!1, b)".
Modified lambda-formal to disallow a number from the idop; otherwise an
ambiguity arises in expressions such as above.  It turns out that
expressions such as "(a!b,c:d)" are perfectly valid, it is the same as
"a ! (b, c): d", in which a is treated as a binding form, b and c are
bound variables, e.g., this is the same as "a(lambda (b,c): d)".

The restriction means we cannot have expressions such as
"lambda 1, x: p(1,x)".  Not much of a restriction, since we can still
use "lambda (1, x): p(1,x)"

This fixes bug # 527


git-svn-id: svn+ssh://spartan.csl.sri.com/svn/public/pvs/trunk@2400 ab98feae-9515-0410-836f-c1e8bd48b6aa