From Jerry James:
Changes to comply with the ANSI Common Lisp specification. This consists
almost entirely of updating the keywords passed to eval-when. The reason I
did this is that I get massive output from SBCL's compiler, and this appears
to be the lowest-hanging fruit in terms of reducing the size of that
output. I'd like to make it smaller, so that it is easier to examine the
output to see if there are any real problems.



git-svn-id: svn+ssh://spartan.csl.sri.com/svn/public/pvs/trunk@5540 ab98feae-9515-0410-836f-c1e8bd48b6aa