From Jerry James:
The Fedora toolchain now embeds a "build ID" in each compiled library and
binary. This is used to uniquely identify debugging information for a given
piece of compiled code. This patch changes the use of build tools slightly
to make sure that a build ID is embedded into each library produced by the
PVS build.



git-svn-id: svn+ssh://spartan.csl.sri.com/svn/public/pvs/trunk@5541 ab98feae-9515-0410-836f-c1e8bd48b6aa