Fix PVS when compiling against SBCL 1.2.2+
SBCL 1.2.2 has a patch that improves handling of the backquote operator.  This
change broke PVS's handling of said operator.  It appears that SBCL no longer
needs special handling of the backquote character, so just disable all such
branches on newer versions of SBCL, and leave a TODO comment for when SBCL
1.2.2+ is the new minimum version.

Also, SBCL never seemed to match the unbackquoted? branch.  So just remove
that code as it is dead anyways (leaving allegro in place).

Finally, bump travis to verify against 1.2.3 as well, since it has the patch.