Better check for proof script changes
Sometimes a proof script is just a sequence of commands that works
interactively, but fails when rerun.  This is because interactively, the
script does not have to respect the proof tree, but on rerun it does.
script-structure-changed? was added to detect such situations, and overwrite
the proof in such cases.