Modified top-lazy-assert-if-with-subst to catch/debug loops
This is called from auto-rewrite*, after it finds a match.
The call to lazy-assert-if-with-subst can recursively call auto-rewrite*
If *tracing-rewrites* is set, prints the rewrite

If the current arguments are in *auto-rewrite-stack*, then treat as a
failure. Otherwise pushes arguments onto *auto-rewrite-stack* and calls
lazy-assert-if-with-subst

In addition, modified simplify methods to bind *auto-rewrite-stack* to nil